Modified getPercentage() method in ColorStackImpl(used streams).

+++ src/main/java/it/unibo/pixArt/model/colorstack/ColorStack.java
@@ -22,7 +22,7 @@ public interface ColorStack {
     /**
      * @return float that rappresents the number of pixels left withing the map as a percentage final_size/init_size
      */
-    float getPercentage();
+    double getPercentage();
 
     /**Method to check if a pixel exists.
      *
